move Distance to own file
